splits controllers, puts models in common, updates ui copy
