+++ admin/app/views/admin.scala.html
@@ -104,7 +104,7 @@
                 <div class="panel-body">
                     <ul>
                         <li><a href="@controllers.admin.routes.CommercialController.renderCommercialMenu()">Commercial Tools</a></li>
-                        <li><a href="@controllers.admin.routes.CommercialController.renderCommercialMenu()">Reader Revenue Tools</a></li>
+                        <li><a href="@controllers.admin.routes.ReaderRevenueAdminController.renderReaderRevenueMenu()">Reader Revenue Tools</a></li>
                     </ul>
                 </div>
             </div>
