+++ admin/app/views/readerRevenue/readerRevenueMenu.scala.html
@@ -10,12 +10,12 @@
     <div class="row">
 
         <div class="col-md-4">
-            <h3>Banner</h3>
+            <h3>Contributions Banner</h3>
             <div class="panel panel-default well--front">
-                <div class="panel-heading">Re-deploy the engagement banner</div>
+                <div class="panel-heading">Tools for contributions (engagement) banner</div>
                 <div class="panel-body">
                     <ul>
-                        <li><a href="@controllers.admin.routes.ReaderRevenueController.renderBannerDeploys()">Banner Deploys</a></li>
+                        <li><a href="@controllers.admin.routes.ReaderRevenueAdminController.renderContributionsBannerAdmin">Redeploy Contributions Banner</a></li>
                     </ul>
                 </div>
             </div>
