+++ applications/conf/routes
@@ -7,8 +7,7 @@ GET         /assets/*path                                                       
 
 GET        /_healthcheck                                                        controllers.HealthCheck.healthCheck()
 
-GET        /reader-revenue/contributions-banner-deploy-log                      controllers.ReaderRevenueController.contributionsBannerDeployLog()
-POST       /reader-revenue/redeploy-contributions-banner                        controllers.ReaderRevenueController.redeployContributionsBanner()
+GET        /reader-revenue/contributions-banner-deploy-log                      controllers.ReaderRevenueAppController.contributionsBannerDeployLog()
 
 GET        /sitemaps/news.xml                                                   controllers.SiteMapController.renderNewsSiteMap()
 GET        /sitemaps/video.xml                                                  controllers.SiteMapController.renderVideoSiteMap()
