+++ common/app/model/readerRevenue/ContributionsBannerDeploy.scala
@@ -1,0 +1,16 @@
+package model.readerRevenue
+
+import org.joda.time.DateTime
+import play.api.libs.json._
+
+
+case class ContributionsBannerDeploy(time: DateTime)
+
+object ContributionsBannerDeploy {
+  private implicit val jodaDateTimeFormats: Format[DateTime] =
+    Format(JodaReads.jodaDateReads("yyyy-MM-dd'T'HH:mm:ssZ"), JodaWrites.jodaDateWrites("yyyy-MM-dd'T'HH:mm:ssZ"))
+  implicit val deployFormat: OFormat[ContributionsBannerDeploy] = Json.format[ContributionsBannerDeploy]
+
+}
+
+
