+++ dev-build/conf/routes
@@ -4,8 +4,11 @@
 
 GET            /commercial/test-page                                                                                             commercial.controllers.CreativeTestPage.allComponents(k: List[String])
 
-GET            /reader-revenue/contributions-banner-deploy-log                                                                   controllers.ReaderRevenueController.contributionsBannerDeployLog()
-POST           /reader-revenue/redeploy-contributions-banner                                                                     controllers.ReaderRevenueController.redeployContributionsBanner()
+# Reader Revenue
+GET            /reader-revenue                                                                                                   controllers.admin.ReaderRevenueAdminController.renderReaderRevenueMenu()
+GET            /reader-revenue/contributions-banner                                                                              controllers.admin.ReaderRevenueAdminController.renderContributionsBannerAdmin()
+POST           /reader-revenue/redeploy-contributions-banner                                                                     controllers.admin.ReaderRevenueAdminController.redeployContributionsBanner()
+GET            /reader-revenue/contributions-banner-deploy-log                                                                   controllers.ReaderRevenueAppController.contributionsBannerDeployLog()
 
 GET            /email-newsletters                                                                                                controllers.SignupPageController.renderNewslettersPage()
